+++ runtimes/eden/src/weights/pallet_scheduler.rs
@@ -50,94 +50,94 @@ impl<T: frame_system::Config> pallet_scheduler::WeightInfo for WeightInfo<T> {
 	// Storage: Scheduler IncompleteSince (r:1 w:1)
 	fn service_agendas_base() -> Weight {
 		// Minimum execution time: 5_131 nanoseconds.
-		Weight::from_ref_time(5_286_000 as u64)
-			.saturating_add(T::DbWeight::get().reads(1 as u64))
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+		Weight::from_ref_time(5_286_000)
+			.saturating_add(T::DbWeight::get().reads(1))
+			.saturating_add(T::DbWeight::get().writes(1))
 	}
 	// Storage: Scheduler Agenda (r:1 w:1)
 	/// The range of component `s` is `[0, 512]`.
 	fn service_agenda_base(s: u32, ) -> Weight {
 		// Minimum execution time: 4_111 nanoseconds.
-		Weight::from_ref_time(8_763_440 as u64)
+		Weight::from_ref_time(8_763_440)
 			// Standard Error: 783
-			.saturating_add(Weight::from_ref_time(372_339 as u64).saturating_mul(s as u64))
-			.saturating_add(T::DbWeight::get().reads(1 as u64))
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+			.saturating_add(Weight::from_ref_time(372_339).saturating_mul(s as u64))
+			.saturating_add(T::DbWeight::get().reads(1))
+			.saturating_add(T::DbWeight::get().writes(1))
 	}
 	fn service_task_base() -> Weight {
 		// Minimum execution time: 10_880 nanoseconds.
-		Weight::from_ref_time(11_194_000 as u64)
+		Weight::from_ref_time(11_194_000)
 	}
 	// Storage: Preimage PreimageFor (r:1 w:1)
 	// Storage: Preimage StatusFor (r:1 w:1)
 	/// The range of component `s` is `[128, 4194304]`.
 	fn service_task_fetched(s: u32, ) -> Weight {
 		// Minimum execution time: 25_347 nanoseconds.
-		Weight::from_ref_time(25_717_000 as u64)
+		Weight::from_ref_time(25_717_000)
 			// Standard Error: 0
-			.saturating_add(Weight::from_ref_time(1_128 as u64).saturating_mul(s as u64))
-			.saturating_add(T::DbWeight::get().reads(2 as u64))
-			.saturating_add(T::DbWeight::get().writes(2 as u64))
+			.saturating_add(Weight::from_ref_time(1_128).saturating_mul(s as u64))
+			.saturating_add(T::DbWeight::get().reads(2))
+			.saturating_add(T::DbWeight::get().writes(2))
 	}
 	// Storage: Scheduler Lookup (r:0 w:1)
 	fn service_task_named() -> Weight {
 		// Minimum execution time: 12_894 nanoseconds.
-		Weight::from_ref_time(13_108_000 as u64)
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+		Weight::from_ref_time(13_108_000)
+			.saturating_add(T::DbWeight::get().writes(1))
 	}
 	fn service_task_periodic() -> Weight {
 		// Minimum execution time: 10_667 nanoseconds.
-		Weight::from_ref_time(10_908_000 as u64)
+		Weight::from_ref_time(10_908_000)
 	}
 	fn execute_dispatch_signed() -> Weight {
 		// Minimum execution time: 4_124 nanoseconds.
-		Weight::from_ref_time(4_680_000 as u64)
+		Weight::from_ref_time(4_680_000)
 	}
 	fn execute_dispatch_unsigned() -> Weight {
 		// Minimum execution time: 4_156 nanoseconds.
-		Weight::from_ref_time(4_361_000 as u64)
+		Weight::from_ref_time(4_361_000)
 	}
 	// Storage: Scheduler Agenda (r:1 w:1)
 	/// The range of component `s` is `[0, 511]`.
 	fn schedule(s: u32, ) -> Weight {
 		// Minimum execution time: 20_504 nanoseconds.
-		Weight::from_ref_time(27_066_818 as u64)
+		Weight::from_ref_time(27_066_818)
 			// Standard Error: 1_114
-			.saturating_add(Weight::from_ref_time(372_897 as u64).saturating_mul(s as u64))
-			.saturating_add(T::DbWeight::get().reads(1 as u64))
-			.saturating_add(T::DbWeight::get().writes(1 as u64))
+			.saturating_add(Weight::from_ref_time(372_897).saturating_mul(s as u64))
+			.saturating_add(T::DbWeight::get().reads(1))
+			.saturating_add(T::DbWeight::get().writes(1))
 	}
 	// Storage: Scheduler Agenda (r:1 w:1)
 	// Storage: Scheduler Lookup (r:0 w:1)
 	/// The range of component `s` is `[1, 512]`.
 	fn cancel(s: u32, ) -> Weight {
 		// Minimum execution time: 21_686 nanoseconds.
-		Weight::from_ref_time(25_696_496 as u64)
+		Weight::from_ref_time(25_696_496)
 			// Standard Error: 1_261
-			.saturating_add(Weight::from_ref_time(362_498 as u64).saturating_mul(s as u64))
-			.saturating_add(T::DbWeight::get().reads(1 as u64))
-			.saturating_add(T::DbWeight::get().writes(2 as u64))
+			.saturating_add(Weight::from_ref_time(362_498).saturating_mul(s as u64))
+			.saturating_add(T::DbWeight::get().reads(1))
+			.saturating_add(T::DbWeight::get().writes(2))
 	}
 	// Storage: Scheduler Lookup (r:1 w:1)
 	// Storage: Scheduler Agenda (r:1 w:1)
 	/// The range of component `s` is `[0, 511]`.
 	fn schedule_named(s: u32, ) -> Weight {
 		// Minimum execution time: 23_084 nanoseconds.
-		Weight::from_ref_time(31_255_518 as u64)
+		Weight::from_ref_time(31_255_518)
 			// Standard Error: 1_258
-			.saturating_add(Weight::from_ref_time(382_534 as u64).saturating_mul(s as u64))
-			.saturating_add(T::DbWeight::get().reads(2 as u64))
-			.saturating_add(T::DbWeight::get().writes(2 as u64))
+			.saturating_add(Weight::from_ref_time(382_534).saturating_mul(s as u64))
+			.saturating_add(T::DbWeight::get().reads(2))
+			.saturating_add(T::DbWeight::get().writes(2))
 	}
 	// Storage: Scheduler Lookup (r:1 w:1)
 	// Storage: Scheduler Agenda (r:1 w:1)
 	/// The range of component `s` is `[1, 512]`.
 	fn cancel_named(s: u32, ) -> Weight {
 		// Minimum execution time: 23_862 nanoseconds.
-		Weight::from_ref_time(28_591_336 as u64)
+		Weight::from_ref_time(28_591_336)
 			// Standard Error: 742
-			.saturating_add(Weight::from_ref_time(369_305 as u64).saturating_mul(s as u64))
-			.saturating_add(T::DbWeight::get().reads(2 as u64))
-			.saturating_add(T::DbWeight::get().writes(2 as u64))
+			.saturating_add(Weight::from_ref_time(369_305).saturating_mul(s as u64))
+			.saturating_add(T::DbWeight::get().reads(2))
+			.saturating_add(T::DbWeight::get().writes(2))
 	}
 }
